Add GitHub stars and Twitter follow badges to readme
In this commit, we've added two new badges to the readme.md file in the starter-kit package. Specifically, a GitHub stars badge and a Twitter follow badge. These are to increase community engagement by encouraging users to star the repository on GitHub and follow developments on Twitter. Also updated the styling of existing badges for consistency.