Add disabler and disable reason in ModelDisablingTrait and update configuration
The functionality of ModelDisablingTrait was expanded to include the concept of a "disabler" and a "disable reason". The method now accepts two parameters - 'reason' and disabler' object that are used to disable the model. The change affects the disabled function and other parts of the script.

Moreover, modifications were made to allow configuration of column names. This allows for better customization and the ability to adapt to different database designs. The changes involved adjustments in StarterKitServiceProvider, adding SK_DISABLER_REQUIRED and SK_DISABLE_REASON_REQUIRED in config file.

New exceptions: DisablerRequiredException and DisableReasonRequiredException were created to handle cases where these new fields are required but not supplied.